+++ streams/ksqldb/statements.replicas-1.sql
@@ -39,7 +39,28 @@ CREATE STREAM NOTIFICATIONS_S (
     REPLICAS=1,
     VALUE_FORMAT = 'JSON');
 
-CREATE STREAM NOTIFICATIONS_INDIVIDUAL_S AS
+CREATE STREAM NOTIFICATIONS_INDIVIDUAL_S (
+    id VARCHAR KEY,
+    subs_id VARCHAR,
+    subs_name VARCHAR,
+    method_id VARCHAR,
+    subs_description VARCHAR,
+    resource_uri VARCHAR,
+    resource_name VARCHAR,
+    resource_description VARCHAR,
+    metric VARCHAR,
+    condition VARCHAR,
+    condition_value VARCHAR,
+    "VALUE" VARCHAR,
+    timestamp VARCHAR,
+    recovery BOOLEAN
+) WITH (
+    KAFKA_TOPIC='NOTIFICATIONS_INDIVIDUAL_S',
+    PARTITIONS=1,
+    REPLICAS=1,
+    VALUE_FORMAT = 'JSON');
+
+INSERT INTO NOTIFICATIONS_INDIVIDUAL_S
 SELECT
     id,
     subs_id,
@@ -278,8 +299,8 @@ SELECT
 FROM SUBS_CONFIG_BLACKBOX_EVENT_S AS s
 PARTITION BY s.owner;
 
-CREATE TABLE SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_T
-(id VARCHAR PRIMARY KEY,
+CREATE TABLE SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_T (
+  id VARCHAR PRIMARY KEY,
   subs_id VARCHAR,
   name VARCHAR,
   description VARCHAR,
@@ -289,15 +310,26 @@ CREATE TABLE SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_T
                   metric VARCHAR,
                   condition VARCHAR,
                   "value" VARCHAR,
-                  "window" BIGINT>,
-)
-WITH (KAFKA_TOPIC='SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_S',
+                  "window" BIGINT>
+) WITH (KAFKA_TOPIC='SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_S',
       PARTITIONS=1,
       REPLICAS=1,
       VALUE_FORMAT='JSON');
 
-CREATE STREAM EVENT_BB_CREATED_S
-AS SELECT
+
+CREATE STREAM EVENT_BB_CREATED_S (
+     id VARCHAR,
+     owner VARCHAR,
+     href VARCHAR,
+     state VARCHAR,
+     timestamp VARCHAR
+) WITH (KAFKA_TOPIC='EVENT_BB_CREATED_S',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+INSERT INTO EVENT_BB_CREATED_S
+SELECT
     e.id as id,
     e.acl->owners[1] as owner,
     e.content->resource->href as href,
@@ -399,7 +431,34 @@ WITH (KAFKA_TOPIC='es_nuvla-nuvlabox-status',
    VALUE_FORMAT='JSON');
 
 -- re-key by parent which is NB ID
-CREATE STREAM NB_TELEM_RESOURCES_REKYED_S AS
+CREATE STREAM NB_TELEM_RESOURCES_REKYED_S (
+   id VARCHAR KEY,
+   name VARCHAR,
+   description VARCHAR,
+   online BOOLEAN,
+   online_prev BOOLEAN,
+   resources STRUCT<cpu STRUCT<"load" DOUBLE, "capacity" BIGINT, "topic" VARCHAR>,
+                    ram STRUCT<"used" BIGINT, "capacity" BIGINT, "topic" VARCHAR>,
+                    disks ARRAY<STRUCT<"used" BIGINT, "capacity" BIGINT, "device" VARCHAR>>>,
+   resources_cpu_load_pers DOUBLE,
+   resources_ram_used_pers DOUBLE,
+   resources_disk1_used_pers DOUBLE,
+   resources_prev STRUCT<cpu STRUCT<"load" DOUBLE, "capacity" BIGINT, "topic" VARCHAR>,
+                           ram STRUCT<"used" BIGINT, "capacity" BIGINT, "topic" VARCHAR>,
+                           disks ARRAY<STRUCT<"used" BIGINT, "capacity" BIGINT, "device" VARCHAR>>>,
+   resources_prev_cpu_load_pers DOUBLE,
+   resources_prev_ram_used_pers DOUBLE,
+   resources_prev_disk1_used_pers DOUBLE,
+   timestamp VARCHAR,
+   acl STRUCT<"owners" ARRAY<VARCHAR>,
+              "view-data" ARRAY<VARCHAR>
+   >
+) WITH (KAFKA_TOPIC='NB_TELEM_RESOURCES_REKYED_S',
+   PARTITIONS=1,
+   REPLICAS=1,
+   VALUE_FORMAT='JSON');
+
+INSERT INTO NB_TELEM_RESOURCES_REKYED_S
 SELECT
     parent as id,
     nb.name as name,
@@ -407,13 +466,13 @@ SELECT
     online,
     "online-prev" as online_prev,
     resources,
-    (resources->cpu->"load" * 100) / resources->cpu->"capacity" as resources_cpu_load_pers,
-    (resources->ram->"used" * 100) / resources->ram->"capacity" as resources_ram_used_pers,
-    (resources->disks[1]->"used" * 100) / resources->disks[1]->"capacity" as resources_disk1_used_pers,
+    CAST((resources->cpu->"load" * 100) / resources->cpu->"capacity" AS DOUBLE) as resources_cpu_load_pers,
+    CAST((resources->ram->"used" * 100) / resources->ram->"capacity" AS DOUBLE) as resources_ram_used_pers,
+    CAST((resources->disks[1]->"used" * 100) / resources->disks[1]->"capacity" AS DOUBLE) as resources_disk1_used_pers,
     "resources-prev" as resources_prev,
-    ("resources-prev"->cpu->"load" * 100) / "resources-prev"->cpu->"capacity" as resources_prev_cpu_load_pers,
-    ("resources-prev"->ram->"used" * 100) / "resources-prev"->ram->"capacity" as resources_prev_ram_used_pers,
-    ("resources-prev"->disks[1]->"used" * 100) / "resources-prev"->disks[1]->"capacity" as resources_prev_disk1_used_pers,
+    CAST(("resources-prev"->cpu->"load" * 100) / "resources-prev"->cpu->"capacity" AS DOUBLE) as resources_prev_cpu_load_pers,
+    CAST(("resources-prev"->ram->"used" * 100) / "resources-prev"->ram->"capacity" AS DOUBLE) as resources_prev_ram_used_pers,
+    CAST(("resources-prev"->disks[1]->"used" * 100) / "resources-prev"->disks[1]->"capacity" AS DOUBLE) as resources_prev_disk1_used_pers,
     "current-time" as timestamp,
     tlm.acl as acl
 FROM NB_TELEM_RESOURCES_S as tlm
@@ -424,8 +483,17 @@ EMIT CHANGES;
 ----------------------------------
 --
 -- Subscription to NB state change
-CREATE TABLE SUBS_NB_STATE_T AS
-SELECT
+CREATE TABLE SUBS_NB_STATE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_STATE_T',
+   PARTITIONS=1,
+   REPLICAS=1,
+   VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_STATE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -468,8 +536,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB load > threshold
-CREATE TABLE SUBS_NB_LOAD_ABOVE_T AS
-SELECT
+CREATE TABLE SUBS_NB_LOAD_ABOVE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_LOAD_ABOVE_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_LOAD_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -516,8 +593,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB load < threshold
-CREATE TABLE SUBS_NB_LOAD_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_LOAD_BELOW_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_LOAD_BELOW_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_LOAD_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -564,8 +650,17 @@ EMIT CHANGES;
 -------------------------------------
 --
 -- Subscription to NB ram > threshold
-CREATE TABLE SUBS_NB_RAM_ABOVE_T AS
-SELECT
+CREATE TABLE SUBS_NB_RAM_ABOVE_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_RAM_ABOVE_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_RAM_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -612,8 +707,17 @@ EMIT CHANGES;
 -------------------------------------
 --
 -- Subscription to NB ram < threshold
-CREATE TABLE SUBS_NB_RAM_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_RAM_BELOW_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_RAM_BELOW_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_RAM_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -659,9 +763,18 @@ EMIT CHANGES;
 
 -------------------------------------
 --
--- Subscription to NB ram > threshold
-CREATE TABLE SUBS_NB_DISK_ABOVE_T AS
-SELECT
+-- Subscription to NB disk > threshold
+CREATE TABLE SUBS_NB_DISK_ABOVE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_DISK_ABOVE_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_DISK_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -708,8 +821,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB disk < threshold
-CREATE TABLE SUBS_NB_DISK_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_DISK_BELOW_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_DISK_BELOW_T',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_DISK_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner