+++ streams/ksqldb/statements.sql
@@ -36,7 +36,28 @@ CREATE STREAM NOTIFICATIONS_S (
     REPLICAS=3,
     VALUE_FORMAT = 'JSON');
 
-CREATE STREAM NOTIFICATIONS_INDIVIDUAL_S AS
+CREATE STREAM NOTIFICATIONS_INDIVIDUAL_S (
+    id VARCHAR KEY,
+    subs_id VARCHAR,
+    subs_name VARCHAR,
+    method_id VARCHAR,
+    subs_description VARCHAR,
+    resource_uri VARCHAR,
+    resource_name VARCHAR,
+    resource_description VARCHAR,
+    metric VARCHAR,
+    condition VARCHAR,
+    condition_value VARCHAR,
+    "VALUE" VARCHAR,
+    timestamp VARCHAR,
+    recovery BOOLEAN
+) WITH (
+    KAFKA_TOPIC='NOTIFICATIONS_INDIVIDUAL_S',
+    PARTITIONS=3,
+    REPLICAS=3,
+    VALUE_FORMAT = 'JSON');
+
+INSERT INTO NOTIFICATIONS_INDIVIDUAL_S
 SELECT
     id,
     subs_id,
@@ -198,6 +219,146 @@ WITH (KAFKA_TOPIC='subscription',
       REPLICAS=3,
       VALUE_FORMAT='JSON');
 
+--
+-- stream with subscription configurations
+CREATE STREAM SUBSCRIPTION_CONFIG_S
+(id VARCHAR KEY,
+  name VARCHAR,
+  description VARCHAR,
+  enabled BOOLEAN,
+  acl STRUCT<
+    owners ARRAY<VARCHAR>
+  >,
+  category VARCHAR,
+  "method-ids" ARRAY<VARCHAR>,
+  "resource-id" VARCHAR,
+  "resource-kind" VARCHAR,
+  criteria STRUCT<kind VARCHAR,
+                  metric VARCHAR,
+                  condition VARCHAR,
+                  "value" VARCHAR,
+                  "window" BIGINT>
+)
+WITH (KAFKA_TOPIC='subscription-config',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+--
+-- stream of events
+CREATE STREAM EVENT_S
+    (id VARCHAR,
+     category VARCHAR,
+     content STRUCT<
+         resource STRUCT<
+             href VARCHAR
+         >,
+         state VARCHAR
+     >,
+     severity VARCHAR,
+     timestamp VARCHAR,
+     tags ARRAY<VARCHAR>,
+     acl STRUCT<
+          owners ARRAY<VARCHAR>
+     >
+    )
+    WITH (KAFKA_TOPIC='event',
+          PARTITIONS=3,
+          REPLICAS=3,
+          VALUE_FORMAT='JSON');
+---    WITH (KAFKA_TOPIC='es_nuvla-event',
+
+------------------------------------------------------
+-- BlackBox creation.
+
+CREATE STREAM SUBS_CONFIG_BLACKBOX_EVENT_S AS
+SELECT
+    s.acl->owners[1] as owner,
+    *
+FROM SUBSCRIPTION_CONFIG_S AS s
+WHERE "resource-kind" = 'event'
+    AND criteria->"value" = 'application/blackbox'
+    AND criteria->metric = 'tag'
+    AND criteria->kind = 'string'
+    AND criteria->condition = 'is'
+EMIT CHANGES;
+
+-- repartitioned by owner
+CREATE STREAM SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_S AS
+SELECT
+    s.owner as id,
+    s.id as subs_id,
+    s.name as name,
+    s.description as description,
+    s.enabled as enabled,
+    s."method-ids" as "method-ids",
+    s.criteria as criteria
+FROM SUBS_CONFIG_BLACKBOX_EVENT_S AS s
+PARTITION BY s.owner;
+
+CREATE TABLE SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_T (
+  id VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  name VARCHAR,
+  description VARCHAR,
+  enabled BOOLEAN,
+  "method-ids" ARRAY<VARCHAR>,
+  criteria STRUCT<kind VARCHAR,
+                  metric VARCHAR,
+                  condition VARCHAR,
+                  "value" VARCHAR,
+                  "window" BIGINT>
+) WITH (KAFKA_TOPIC='SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_S',
+      PARTITIONS=1,
+      REPLICAS=1,
+      VALUE_FORMAT='JSON');
+
+
+CREATE STREAM EVENT_BB_CREATED_S (
+     id VARCHAR,
+     owner VARCHAR,
+     href VARCHAR,
+     state VARCHAR,
+     timestamp VARCHAR
+) WITH (KAFKA_TOPIC='EVENT_BB_CREATED_S',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+INSERT INTO EVENT_BB_CREATED_S
+SELECT
+    e.id as id,
+    e.acl->owners[1] as owner,
+    e.content->resource->href as href,
+    e.content->state as state,
+    e.timestamp as timestamp
+FROM EVENT_S as e
+WHERE e.category = 'user'
+    AND ARRAY_CONTAINS(e.tags, 'application/blackbox')
+    AND e.content->resource->href LIKE 'data-record/%'
+    AND e.content->state = 'created'
+EMIT CHANGES;
+
+INSERT INTO NOTIFICATIONS_S
+SELECT
+     subs_t.id as id,
+     AS_VALUE(subs_t.subs_id) as subs_id,
+     subs_t.name as subs_name,
+     subs_t."method-ids" as method_ids,
+     subs_t.description as subs_description,
+     CONCAT('api/', bbe.href) as resource_uri,
+     'blackbox' as resource_name,
+     'blackbox' as resource_description,
+     'content-type' as metric,
+     subs_t.criteria->condition as condition,
+     CAST(subs_t.criteria->"value" as VARCHAR) as condition_value,
+     'true' as "VALUE",
+     CONCAT(SPLIT(bbe.timestamp, '.')[1], 'Z') as timestamp,
+     true as recovery
+FROM EVENT_BB_CREATED_S AS bbe
+JOIN SUBS_CONFIG_BLACKBOX_EVENT_BY_OWNER_T AS subs_t ON bbe.owner = subs_t.id
+EMIT CHANGES;
+
 ---------------------------------
 -- NuvlaBox.
 
@@ -267,8 +428,34 @@ WITH (KAFKA_TOPIC='es_nuvla-nuvlabox-status',
    VALUE_FORMAT='JSON');
 
 -- re-key by parent which is NB ID
--- due to: only T-to-T join on foreign key is supported.
-CREATE STREAM NB_TELEM_RESOURCES_REKYED_S AS
+CREATE STREAM NB_TELEM_RESOURCES_REKYED_S (
+   id VARCHAR KEY,
+   name VARCHAR,
+   description VARCHAR,
+   online BOOLEAN,
+   online_prev BOOLEAN,
+   resources STRUCT<cpu STRUCT<"load" DOUBLE, "capacity" BIGINT, "topic" VARCHAR>,
+                    ram STRUCT<"used" BIGINT, "capacity" BIGINT, "topic" VARCHAR>,
+                    disks ARRAY<STRUCT<"used" BIGINT, "capacity" BIGINT, "device" VARCHAR>>>,
+   resources_cpu_load_pers DOUBLE,
+   resources_ram_used_pers DOUBLE,
+   resources_disk1_used_pers DOUBLE,
+   resources_prev STRUCT<cpu STRUCT<"load" DOUBLE, "capacity" BIGINT, "topic" VARCHAR>,
+                           ram STRUCT<"used" BIGINT, "capacity" BIGINT, "topic" VARCHAR>,
+                           disks ARRAY<STRUCT<"used" BIGINT, "capacity" BIGINT, "device" VARCHAR>>>,
+   resources_prev_cpu_load_pers DOUBLE,
+   resources_prev_ram_used_pers DOUBLE,
+   resources_prev_disk1_used_pers DOUBLE,
+   timestamp VARCHAR,
+   acl STRUCT<"owners" ARRAY<VARCHAR>,
+              "view-data" ARRAY<VARCHAR>
+   >
+) WITH (KAFKA_TOPIC='NB_TELEM_RESOURCES_REKYED_S',
+   PARTITIONS=3,
+   REPLICAS=3,
+   VALUE_FORMAT='JSON');
+
+INSERT INTO NB_TELEM_RESOURCES_REKYED_S
 SELECT
     parent as id,
     nb.name as name,
@@ -276,13 +463,13 @@ SELECT
     online,
     "online-prev" as online_prev,
     resources,
-    (resources->cpu->"load" * 100) / resources->cpu->"capacity" as resources_cpu_load_pers,
-    (resources->ram->"used" * 100) / resources->ram->"capacity" as resources_ram_used_pers,
-    (resources->disks[1]->"used" * 100) / resources->disks[1]->"capacity" as resources_disk1_used_pers,
+    CAST((resources->cpu->"load" * 100) / resources->cpu->"capacity" AS DOUBLE) as resources_cpu_load_pers,
+    CAST((resources->ram->"used" * 100) / resources->ram->"capacity" AS DOUBLE) as resources_ram_used_pers,
+    CAST((resources->disks[1]->"used" * 100) / resources->disks[1]->"capacity" AS DOUBLE) as resources_disk1_used_pers,
     "resources-prev" as resources_prev,
-    ("resources-prev"->cpu->"load" * 100) / "resources-prev"->cpu->"capacity" as resources_prev_cpu_load_pers,
-    ("resources-prev"->ram->"used" * 100) / "resources-prev"->ram->"capacity" as resources_prev_ram_used_pers,
-    ("resources-prev"->disks[1]->"used" * 100) / "resources-prev"->disks[1]->"capacity" as resources_prev_disk1_used_pers,
+    CAST(("resources-prev"->cpu->"load" * 100) / "resources-prev"->cpu->"capacity" AS DOUBLE) as resources_prev_cpu_load_pers,
+    CAST(("resources-prev"->ram->"used" * 100) / "resources-prev"->ram->"capacity" AS DOUBLE) as resources_prev_ram_used_pers,
+    CAST(("resources-prev"->disks[1]->"used" * 100) / "resources-prev"->disks[1]->"capacity" AS DOUBLE) as resources_prev_disk1_used_pers,
     "current-time" as timestamp,
     tlm.acl as acl
 FROM NB_TELEM_RESOURCES_S as tlm
@@ -293,8 +480,17 @@ EMIT CHANGES;
 ----------------------------------
 --
 -- Subscription to NB state change
-CREATE TABLE SUBS_NB_STATE_T AS
-SELECT
+CREATE TABLE SUBS_NB_STATE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_STATE_T',
+   PARTITIONS=3,
+   REPLICAS=3,
+   VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_STATE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -337,8 +533,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB load > threshold
-CREATE TABLE SUBS_NB_LOAD_ABOVE_T AS
-SELECT
+CREATE TABLE SUBS_NB_LOAD_ABOVE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_LOAD_ABOVE_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_LOAD_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -385,8 +590,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB load < threshold
-CREATE TABLE SUBS_NB_LOAD_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_LOAD_BELOW_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_LOAD_BELOW_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_LOAD_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -433,8 +647,17 @@ EMIT CHANGES;
 -------------------------------------
 --
 -- Subscription to NB ram > threshold
-CREATE TABLE SUBS_NB_RAM_ABOVE_T AS
-SELECT
+CREATE TABLE SUBS_NB_RAM_ABOVE_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_RAM_ABOVE_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_RAM_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -481,8 +704,17 @@ EMIT CHANGES;
 -------------------------------------
 --
 -- Subscription to NB ram < threshold
-CREATE TABLE SUBS_NB_RAM_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_RAM_BELOW_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_RAM_BELOW_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_RAM_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -528,9 +760,18 @@ EMIT CHANGES;
 
 -------------------------------------
 --
--- Subscription to NB ram > threshold
-CREATE TABLE SUBS_NB_DISK_ABOVE_T AS
-SELECT
+-- Subscription to NB disk > threshold
+CREATE TABLE SUBS_NB_DISK_ABOVE_T (
+ "resource-id" VARCHAR PRIMARY KEY,
+ subs_id VARCHAR,
+ owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_DISK_ABOVE_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_DISK_ABOVE_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner
@@ -577,8 +818,17 @@ EMIT CHANGES;
 --------------------------------------
 --
 -- Subscription to NB disk < threshold
-CREATE TABLE SUBS_NB_DISK_BELOW_T AS
-SELECT
+CREATE TABLE SUBS_NB_DISK_BELOW_T (
+  "resource-id" VARCHAR PRIMARY KEY,
+  subs_id VARCHAR,
+  owner VARCHAR
+) WITH (KAFKA_TOPIC='SUBS_NB_DISK_BELOW_T',
+      PARTITIONS=3,
+      REPLICAS=3,
+      VALUE_FORMAT='JSON');
+
+CREATE OR REPLACE TABLE SUBS_NB_DISK_BELOW_T
+AS SELECT
     "resource-id",
     LATEST_BY_OFFSET(AS_VALUE(id)) AS subs_id,
     LATEST_BY_OFFSET(acl->owners[1]) AS owner